+++ test/OpenTelemetry.AutoInstrumentation.Tests/Configurations/SettingsTests.cs
@@ -58,6 +58,7 @@ public class SettingsTests : IDisposable
 
             // Instrumentation options tests
             settings.InstrumentationOptions.GraphQLSetDocument.Should().BeFalse();
+            settings.InstrumentationOptions.OracleMdaSetDbStatementForText.Should().BeFalse();
             settings.InstrumentationOptions.SqlClientSetDbStatementForText.Should().BeFalse();
 #if NET6_0_OR_GREATER
             settings.InstrumentationOptions.EntityFrameworkCoreSetDbStatementForText.Should().BeFalse();