+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
   and [Oracle.ManagedDataAccess](https://www.nuget.org/packages/Oracle.ManagedDataAccess)
   traces instrumentation from 23.3.2-dev together with support for
   `OTEL_DOTNET_AUTO_ORACLEMDA_SET_DBSTATEMENT_FOR_TEXT` environment variable.
+  ARM64 platform is not supported.
 
 ### Changed