+++ CHANGELOG.md
@@ -34,7 +34,7 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 ### Fixed
 
-* Resolved a crash issue caused by `System.IO.FileLoadException` for
+- Resolved a crash issue caused by `System.IO.FileLoadException` for
   `Microsoft.Extensions*.dll` libraries. This issue was due to a conflict with
   runtime store libraries, impacting applications with mismatched dependency
   versions. This fix enhances stability by addressing the underlying